new feature : golden ticket
New feature implemented for Devoxx UK and FR. Attendees can buy one of the 64 Golden Ticket. A GT holder gets a special access to the CFP, can review and evalutes the Proposals. However a GT cannot see the private comments between the CFP program committee and the speaker. Other notes are not shown.