feat(emails): notification to program committee can now be unactivate from configuration
For France we receive 200 emails per day during the hot CFP weeks. And this is useless. The way we work is that when a new proposal gets submitted, it's added to our TODO list.

I'd like to introduce a Cron task that would send the todo list size to each CFP menber. Something like "Nicolas you have 600 talks to review"