bug(speaker): Webuser visitor to speaker did not work
When a Visitor or a GT wants to be also a Speaker, some code in CFPAdmin tries to import and to create the Speaker. However, it didn’t add the UUID to the Webuser:speaker SET.

So this was generating bugs. One could not add a secondary speaker for instance.

This has been fixed