feat(publisher): Check published Agenda and update slots
When a speaker edits one of its proposals but the Agenda was already published, we’d like to review and to confirm (or reject) an update. A published schedule contains Slots, with a copy of the Proposal. That’s because we don’t want speakers to live update their title or abstract once the agenda is published.